MCSC-354 Content updated

+++ app/pages/annualallowance/taxyear/TotalIncomePage.scala
@@ -21,7 +21,7 @@ import pages.QuestionPage
 import play.api.libs.json.JsPath
 import play.api.mvc.Call
 
-case class TotalIncomePage(period: Period, schemeIndex: SchemeIndex) extends QuestionPage[Int] {
+case class TotalIncomePage(period: Period, schemeIndex: SchemeIndex) extends QuestionPage[BigInt] {
 
   override def path: JsPath = JsPath \ "aa" \ "years" \ period.toString \ "schemes" \ schemeIndex.toString \ toString
